requirements.txt
discord.py
youtube_dl
youtube-search-python
lyricsgenius~=3.0.1
psutil~=5.9.8
discord-pretty-help
pynacl
numpy~=1.26.4
matplotlib~=3.8.3
scipy~=1.12.0
jishaku